Create AbstractArrayStorage.java, define template method getIndex in it.
